requirements.txt
python-telegram-bot==21.1.1
python-dotenv==1.0.1
tabulate==0.9.0
openpyxl==3.1.2
apscheduler==3.10.1
